+++ include/H2DE/H2DE_engine.h
@@ -113,6 +113,8 @@ public:
     friend void H2DE_ResumeSfx(const H2DE_Engine* engine, int id);
     friend void H2DE_SetSongVolume(const H2DE_Engine* engine, int volume);
     friend void H2DE_SetSfxVolume(const H2DE_Engine* engine, int volume);
+    friend bool H2DE_IsSongPlaying(const H2DE_Engine* engine);
+    friend bool H2DE_IsSfxPlaying(const H2DE_Engine* engine, int id);
 
     friend unsigned int H2DE_Delay(const H2DE_Engine* engine, unsigned int ms, const std::function<void()>& callback, bool pauseSensitive);
     friend void H2DE_ResetDelay(const H2DE_Engine* engine, unsigned int id);